Avoid touching the global strategy
When the global strategy object was modified, there was a race condition
if the server was processing multiple requests targeting different badges.

The solution implemented in this commit doesn't anymore modify the strategy,
but instead, passes the right consumer values in the request object and to
the passport.authenticate function.

This functionality depends on a forked passport-oauth2.